\name{peresneto}
\alias{peresneto}
\docType{data}
\title{Test correlation matrices}
\description{
  Test correlation matrices used in papers by Pedro R. Peres-Neto to test methods for assessing principal components analyses.
}
\usage{data(peresneto)}
\format{
  The format is:
Two lists of 18 correlation matrices
 
}
\source{
  Peres-Neto et al. (2005)
}
\references{
  P.R. Peres-Neto, D.A. Jackson and K.M. Somers (2005). How many principal components? stopping rules for determining the number of non=trivial axes revisited.  Computational Statistics and Data Analysis. 49: 974-997.
}
\examples{
data(peresneto)
}
\keyword{datasets}
