\name{reo-package}
\alias{reo-package}
\alias{reo}
\docType{package}
\title{
Random effects ordination
}
\description{
Functions for fitting ordination models with random effects.
}
\details{
\tabular{ll}{
Package: \tab reo\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-01-20\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
The most important functions are \code{\link{factanal.predictive}} and \code{\link{ltm.ecol}}.
}
\references{
Walker, S.C. and Jackson, D.A. (in press) Random-effects ordination: describing and predicting multivariate correlations and co-occurrences.  Ecological Monographs. 
}
\author{
Steve Walker
}
\keyword{ package }
